local_server/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-telegram-bot==20.8
aiohttp==3.9.3
python-dotenv==1.0.1 